requirements.txt
PyYAML>=5.1
click>=7.0
requests>=2.25.0
packaging>=20.0
pytest>=6.0  # For testing